+++ core-backend/common/src/memory.rs
@@ -129,9 +129,9 @@ pub trait MemoryOwner {
 /// ```
 #[derive(Debug)]
 pub struct MemoryAccessManager<E> {
-    reads: Vec<MemoryInterval>,
-    writes: Vec<MemoryInterval>,
-    _phantom: PhantomData<E>,
+    pub(crate) reads: Vec<MemoryInterval>,
+    pub(crate) writes: Vec<MemoryInterval>,
+    pub(crate) _phantom: PhantomData<E>,
 }
 
 impl<E> Default for MemoryAccessManager<E> {
@@ -194,7 +194,7 @@ impl<E> MemoryAccessRecorder for MemoryAccessManager<E> {
 
 impl<E: BackendExt> MemoryAccessManager<E> {
     /// Call pre-processing of registered memory accesses. Clear `self.reads` and `self.writes`.
-    fn pre_process_memory_accesses(
+    pub(crate) fn pre_process_memory_accesses(
         &mut self,
         gas_left: &mut GasLeft,
     ) -> Result<(), MemoryAccessError> {
@@ -211,7 +211,7 @@ impl<E: BackendExt> MemoryAccessManager<E> {
     }
 
     /// Pre-process registered accesses if need and read data from `memory` to `buff`.
-    fn read_into_buf<M: Memory>(
+    pub(crate) fn read_into_buf<M: Memory>(
         &mut self,
         memory: &M,
         ptr: u32,
@@ -338,563 +338,30 @@ fn read_memory_as<T: Sized>(memory: &impl Memory, ptr: u32) -> Result<T, MemoryE
 
 /// Read static size type access wrapper.
 pub struct WasmMemoryReadAs<T> {
-    ptr: u32,
-    _phantom: PhantomData<T>,
+    pub(crate) ptr: u32,
+    pub(crate) _phantom: PhantomData<T>,
 }
 
 /// Read decoded type access wrapper.
 pub struct WasmMemoryReadDecoded<T: Decode + MaxEncodedLen> {
-    ptr: u32,
-    _phantom: PhantomData<T>,
+    pub(crate) ptr: u32,
+    pub(crate) _phantom: PhantomData<T>,
 }
 
 /// Read access wrapper.
 pub struct WasmMemoryRead {
-    ptr: u32,
-    size: u32,
+    pub(crate) ptr: u32,
+    pub(crate) size: u32,
 }
 
 /// Write static size type access wrapper.
 pub struct WasmMemoryWriteAs<T> {
-    ptr: u32,
-    _phantom: PhantomData<T>,
+    pub(crate) ptr: u32,
+    pub(crate) _phantom: PhantomData<T>,
 }
 
 /// Write access wrapper.
 pub struct WasmMemoryWrite {
-    ptr: u32,
-    size: u32,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::mock::{MockExt, MockMemory};
-    use gear_core::memory::WASM_PAGE_SIZE;
-
-    use scale_info::scale::{self, Decode, Encode};
-
-    const GAS_LEFT: GasLeft = GasLeft {
-        gas: core::u64::MAX,
-        allowance: core::u64::MAX,
-    };
-
-    #[test]
-    fn test_pre_process_memory_accesses_with_no_accesses() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.pre_process_memory_accesses(&mut gas_left);
-
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_pre_process_memory_accesses_with_only_reads() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_read(0, 10);
-
-        let result = memory_access_manager.pre_process_memory_accesses(&mut gas_left);
-
-        assert!(result.is_ok());
-        assert!(memory_access_manager.reads.is_empty());
-    }
-
-    #[test]
-    fn test_pre_process_memory_accesses_with_only_writes() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_write(0, 10);
-
-        let result = memory_access_manager.pre_process_memory_accesses(&mut gas_left);
-
-        assert!(result.is_ok());
-        assert!(memory_access_manager.writes.is_empty());
-    }
-
-    #[test]
-    fn test_pre_process_memory_accesses_with_reads_and_writes() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_read(0, 10);
-        memory_access_manager.register_write(10, 20);
-
-        let result = memory_access_manager.pre_process_memory_accesses(&mut gas_left);
-
-        assert!(result.is_ok());
-        assert!(memory_access_manager.reads.is_empty());
-        assert!(memory_access_manager.writes.is_empty());
-    }
-
-    #[test]
-    fn test_read_into_buf_with_zero_size_buffer() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result =
-            memory_access_manager.read_into_buf(&MockMemory::new(1), 0, &mut [], &mut gas_left);
-
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_read_into_buf_with_same_size_buffer() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_read(0, 10);
-
-        let result = memory_access_manager.read_into_buf(
-            &MockMemory::new(1),
-            0,
-            &mut [0u8; 10],
-            &mut gas_left,
-        );
-
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_read_with_valid_memory_access() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_read(0, 10);
-
-        let memory = &mut MockMemory::new(1);
-        memory.write(0, &[5u8; 10]).unwrap();
-
-        let result =
-            memory_access_manager.read(memory, WasmMemoryRead { ptr: 0, size: 10 }, &mut gas_left);
-
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), &[5u8; 10]);
-    }
-
-    #[test]
-    fn test_read_with_empty_memory_access() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.read(
-            &MockMemory::new(10),
-            WasmMemoryRead { ptr: 0, size: 0 },
-            &mut gas_left,
-        );
-
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_empty());
-    }
-
-    #[test]
-    fn test_read_decoded_with_valid_encoded_data() {
-        #[derive(Encode, Decode, Debug, PartialEq)]
-        #[codec(crate = scale)]
-        struct MockEncodeData {
-            data: u64,
-        }
-
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_read_decoded::<u64>(0);
-
-        let memory = &mut MockMemory::new(1);
-        let encoded = MockEncodeData { data: 1234 }.encode();
-        memory.write(0, &encoded).unwrap();
-
-        let result = memory_access_manager.read_decoded::<MockMemory, u64>(
-            memory,
-            WasmMemoryReadDecoded {
-                ptr: 0,
-                _phantom: PhantomData,
-            },
-            &mut gas_left,
-        );
-
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), 1234u64);
-    }
-
-    #[test]
-    fn test_read_decoded_with_invalid_encoded_data() {
-        #[derive(Debug)]
-        struct InvalidDecode {}
-
-        impl Decode for InvalidDecode {
-            fn decode<T>(_input: &mut T) -> Result<Self, scale_info::scale::Error> {
-                Err("Invalid decoding".into())
-            }
-        }
-
-        impl Encode for InvalidDecode {
-            fn encode_to<T: scale_info::scale::Output + ?Sized>(&self, _dest: &mut T) {}
-        }
-
-        impl MaxEncodedLen for InvalidDecode {
-            fn max_encoded_len() -> usize {
-                0
-            }
-        }
-
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_read_decoded::<InvalidDecode>(0);
-
-        let memory = &mut MockMemory::new(1);
-        let encoded = alloc::vec![7u8; gear_core::memory::WASM_PAGE_SIZE];
-        memory.write(0, &encoded).unwrap();
-
-        let result = memory_access_manager.read_decoded::<MockMemory, InvalidDecode>(
-            &MockMemory::new(1),
-            WasmMemoryReadDecoded {
-                ptr: 0,
-                _phantom: PhantomData,
-            },
-            &mut gas_left,
-        );
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_read_decoded_reading_error() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_read_decoded::<u64>(0);
-
-        let result = memory_access_manager.read_decoded::<MockMemory, u64>(
-            &MockMemory::new(1),
-            WasmMemoryReadDecoded {
-                ptr: u32::MAX,
-                _phantom: PhantomData,
-            },
-            &mut gas_left,
-        );
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_read_as_with_valid_data() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_read_as::<u64>(0);
-
-        let memory = &mut MockMemory::new(1);
-        let encoded = 1234u64.to_le_bytes();
-        memory.write(0, &encoded).unwrap();
-
-        let result = memory_access_manager.read_as::<MockMemory, u64>(
-            memory,
-            WasmMemoryReadAs {
-                ptr: 0,
-                _phantom: PhantomData,
-            },
-            &mut gas_left,
-        );
-
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), 1234);
-    }
-
-    #[test]
-    fn test_read_as_with_invalid_pointer() {
-        let mut gas_left = GAS_LEFT;
-        let memory = &mut MockMemory::new(1);
-
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_read_as::<u64>(0);
-
-        let result = memory_access_manager.read_as::<MockMemory, u128>(
-            memory,
-            WasmMemoryReadAs {
-                ptr: u32::MAX,
-                _phantom: PhantomData,
-            },
-            &mut gas_left,
-        );
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_write_with_zero_size_interval() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_write(0, 0);
-
-        let result = memory_access_manager.write(
-            &mut MockMemory::new(1),
-            WasmMemoryWrite { ptr: 0, size: 0 },
-            &[],
-            &mut gas_left,
-        );
-
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_write_with_zero_buffer_size() {
-        extern crate std;
-
-        let result = std::panic::catch_unwind(|| {
-            let mut gas_left = GAS_LEFT;
-            let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-            memory_access_manager.register_write(0, 10);
-
-            memory_access_manager.write(
-                &mut MockMemory::new(1),
-                WasmMemoryWrite { ptr: 0, size: 10 },
-                &[],
-                &mut gas_left,
-            )
-        });
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_write_with_same_buffer_size() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_write(0, 10);
-        let buffer = [0u8; 10];
-
-        let result = memory_access_manager.write(
-            &mut MockMemory::new(1),
-            WasmMemoryWrite { ptr: 0, size: 10 },
-            &buffer,
-            &mut gas_left,
-        );
-
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_write_with_larger_buffer_size() {
-        extern crate std;
-
-        let result = std::panic::catch_unwind(|| {
-            let mut gas_left = GAS_LEFT;
-            let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-            memory_access_manager.register_write(0, 10);
-            let buffer = [0u8; 20];
-
-            memory_access_manager.write(
-                &mut MockMemory::new(1),
-                WasmMemoryWrite { ptr: 0, size: 10 },
-                &buffer,
-                &mut gas_left,
-            )
-        });
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_write_as_with_zero_size_object() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_write_as::<u32>(0);
-
-        let result = memory_access_manager.write_as(
-            &mut MockMemory::new(1),
-            WasmMemoryWriteAs::<u32> {
-                ptr: 0,
-                _phantom: PhantomData,
-            },
-            0,
-            &mut gas_left,
-        );
-
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_write_as_with_same_object_size() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_write_as::<u8>(0);
-
-        let result = memory_access_manager.write_as(
-            &mut MockMemory::new(1),
-            WasmMemoryWriteAs {
-                ptr: 0,
-                _phantom: PhantomData,
-            },
-            1u8,
-            &mut gas_left,
-        );
-
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_write_as_with_larger_object_size() {
-        let mut gas_left = GAS_LEFT;
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-        memory_access_manager.register_write_as::<u8>(0);
-
-        let result = memory_access_manager.write_as(
-            &mut MockMemory::new(1),
-            WasmMemoryWriteAs {
-                ptr: WASM_PAGE_SIZE as u32,
-                _phantom: PhantomData,
-            },
-            7u8,
-            &mut gas_left,
-        );
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_register_read_with_valid_interval() {
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.register_read(0, 10);
-
-        assert_eq!(result.ptr, 0);
-        assert_eq!(result.size, 10);
-        assert_eq!(memory_access_manager.reads.len(), 1);
-        assert_eq!(memory_access_manager.writes.len(), 0);
-    }
-
-    #[test]
-    fn test_register_read_with_zero_size() {
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.register_read(0, 0);
-
-        assert_eq!(result.ptr, 0);
-        assert_eq!(result.size, 0);
-        assert_eq!(memory_access_manager.reads.len(), 1);
-        assert_eq!(memory_access_manager.writes.len(), 0);
-    }
-
-    #[test]
-    fn test_register_read_as_with_valid_interval() {
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.register_read_as::<u8>(0);
-
-        assert_eq!(result.ptr, 0);
-        assert_eq!(memory_access_manager.reads.len(), 1);
-        assert_eq!(memory_access_manager.writes.len(), 0);
-        assert_eq!(memory_access_manager.reads[0].offset, 0);
-        assert_eq!(
-            memory_access_manager.reads[0].size,
-            core::mem::size_of::<u8>() as u32
-        );
-    }
-
-    #[test]
-    fn test_register_read_as_with_zero_size() {
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.register_read_as::<u8>(0);
-
-        assert_eq!(result.ptr, 0);
-        assert_eq!(memory_access_manager.reads.len(), 1);
-        assert_eq!(memory_access_manager.writes.len(), 0);
-        assert_eq!(memory_access_manager.reads[0].offset, 0);
-        assert_eq!(
-            memory_access_manager.reads[0].size,
-            core::mem::size_of::<u8>() as u32
-        );
-    }
-
-    #[derive(Debug, PartialEq, Eq, Encode, Decode, MaxEncodedLen)]
-    #[codec(crate = scale)]
-    struct TestStruct {
-        a: u32,
-        b: u64,
-    }
-
-    #[test]
-    fn test_register_read_decoded_with_valid_interval() {
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.register_read_decoded::<TestStruct>(0);
-
-        assert_eq!(result.ptr, 0);
-        assert_eq!(memory_access_manager.reads.len(), 1);
-        assert_eq!(memory_access_manager.writes.len(), 0);
-        assert_eq!(memory_access_manager.reads[0].offset, 0);
-        assert_eq!(
-            memory_access_manager.reads[0].size,
-            TestStruct::max_encoded_len() as u32
-        );
-    }
-
-    #[test]
-    fn test_register_read_decoded_with_zero_size() {
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.register_read_decoded::<TestStruct>(0);
-
-        assert_eq!(result.ptr, 0);
-        assert_eq!(memory_access_manager.reads.len(), 1);
-        assert_eq!(memory_access_manager.writes.len(), 0);
-        assert_eq!(memory_access_manager.reads[0].offset, 0);
-        assert_eq!(
-            memory_access_manager.reads[0].size,
-            TestStruct::max_encoded_len() as u32
-        );
-    }
-
-    #[test]
-    fn test_register_write_with_valid_interval() {
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.register_write(0, 10);
-
-        assert_eq!(result.ptr, 0);
-        assert_eq!(result.size, 10);
-        assert_eq!(memory_access_manager.reads.len(), 0);
-        assert_eq!(memory_access_manager.writes.len(), 1);
-    }
-
-    #[test]
-    fn test_register_write_with_zero_size() {
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.register_write(0, 0);
-
-        assert_eq!(result.ptr, 0);
-        assert_eq!(result.size, 0);
-        assert_eq!(memory_access_manager.reads.len(), 0);
-        assert_eq!(memory_access_manager.writes.len(), 1);
-    }
-
-    #[test]
-    fn test_register_write_as_with_valid_interval() {
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.register_write_as::<u8>(0);
-
-        assert_eq!(result.ptr, 0);
-        assert_eq!(memory_access_manager.reads.len(), 0);
-        assert_eq!(memory_access_manager.writes.len(), 1);
-        assert_eq!(memory_access_manager.writes[0].offset, 0);
-        assert_eq!(
-            memory_access_manager.writes[0].size,
-            core::mem::size_of::<u8>() as u32
-        );
-    }
-
-    #[test]
-    fn test_register_write_as_with_zero_size() {
-        let mut memory_access_manager = MemoryAccessManager::<MockExt>::default();
-
-        let result = memory_access_manager.register_write_as::<u8>(0);
-
-        assert_eq!(result.ptr, 0);
-        assert_eq!(memory_access_manager.reads.len(), 0);
-        assert_eq!(memory_access_manager.writes.len(), 1);
-        assert_eq!(memory_access_manager.writes[0].offset, 0);
-        assert_eq!(
-            memory_access_manager.writes[0].size,
-            core::mem::size_of::<u8>() as u32
-        );
-    }
+    pub(crate) ptr: u32,
+    pub(crate) size: u32,
 }
